+++ gui/MainWindowStatusBar.py
@@ -1,0 +1,56 @@
+"""Just a status bar with a custom separator"""
+
+from PyQt6.QtWidgets import QStatusBar, QLabel, QFrame
+from inputremapper.gui.components.main import StatusBar
+
+
+class MainWindowStatusBar(QStatusBar):
+
+    def __init__(self, parent=None):
+        super(MainWindowStatusBar, self).__init__(parent)
+
+        self.setSizeGripEnabled(False)
+
+        self._selection_length_info_label = QLabel()
+        self._selection_info_separator = StatusBarSeparator()
+        self._line_col_info_label = QLabel()
+        self._line_col_info_separator = StatusBarSeparator()
+        self._installation_path_label = QLabel("No PyWright folder selected")
+
+        self.addPermanentWidget(self._selection_length_info_label)
+        self.addPermanentWidget(self._selection_info_separator)
+        self.addPermanentWidget(self._line_col_info_label)
+        self.addPermanentWidget(self._line_col_info_separator)
+        self.addPermanentWidget(self._installation_path_label)
+
+        self.setContentsMargins(4, 2, 4, 2)
+
+    def set_selection_length_info(self, selection_length: int):
+        if selection_length > 0:
+            self._selection_info_separator.setHidden(False)
+            self._selection_length_info_label.setText("{} characters selected".format(selection_length))
+        else:
+            self._selection_info_separator.setHidden(True)
+            self._selection_length_info_label.setText("")
+
+    def set_installation_path_info(self, installation_path: str):
+        self._installation_path_label.setText(installation_path)
+
+    def set_cursor_position_info(self, line: int, col: int):
+        """Displays cursor position in a line/column manner.
+        Will show nothing when either line or col is non-positive.
+        :param line: Line number
+        :param col: Column number"""
+        if line > 0 and col > 0:
+            self._line_col_info_separator.setHidden(False)
+            self._line_col_info_label.setText("Ln: {}, Col: {}".format(line, col))
+        else:
+            self._line_col_info_separator.setHidden(True)
+            self._line_col_info_label.setText("")
+
+
+class StatusBarSeparator(QFrame):
+    """Literally just a vertical line. Why QStatusBar simply doesn't have an addSeparator() method is anyone's guess."""
+    def __init__(self, parent=None):
+        super().__init__()
+        self.setFrameStyle(QFrame.Shape.VLine)